Added notifications content view with sample data
