+++ src/Smallify.Module.Notifications/Samples/NotificationsShellViewModelData.cs
@@ -1,5 +1,4 @@
 ﻿using Smallify.Module.Notifications.ViewModels;
-using System.Collections.ObjectModel;
 using System.Windows.Input;
 
 namespace Smallify.Module.Notifications.Samples
@@ -8,19 +7,6 @@ namespace Smallify.Module.Notifications.Samples
 	{
 		public ICommand ExitCommand => throw new System.NotImplementedException();
 
-		public ObservableCollection<string> Notifications
-		{
-			get => new ObservableCollection<string>()
-			{
-				"Fake Notification #1",
-				"Fake Notitication #2",
-				"Fake Notitication #3",
-				"Fake Notitication #4",
-				"Fake Notitication #5"
-			};
-			set => throw new System.NotImplementedException();
-		}
-
-
+		public INotificationsViewModel NotificationsViewModel => new NotificationsViewModelData();
 	}
 }
