+++ src/Smallify.Module.Notifications/Smallify.Module.Notifications.csproj
@@ -73,15 +73,21 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Samples\NotificationsButtonViewModelData.cs" />
     <Compile Include="Samples\NotificationsShellViewModelData.cs" />
+    <Compile Include="Samples\NotificationsViewModelData.cs" />
     <Compile Include="ViewModels\INotificationsButtonViewModel.cs" />
     <Compile Include="ViewModels\INotificationsShellViewModel.cs" />
+    <Compile Include="ViewModels\INotificationsViewModel.cs" />
     <Compile Include="ViewModels\NotificationsButtonViewModel.cs" />
     <Compile Include="ViewModels\NotificationsShellViewModel.cs" />
+    <Compile Include="ViewModels\NotificationsViewModel.cs" />
     <Compile Include="Views\NotificationsButtonView.xaml.cs">
       <DependentUpon>NotificationsButtonView.xaml</DependentUpon>
     </Compile>
     <Compile Include="Views\NotificationsShell.xaml.cs">
       <DependentUpon>NotificationsShell.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="Views\NotificationsView.xaml.cs">
+      <DependentUpon>NotificationsView.xaml</DependentUpon>
     </Compile>
   </ItemGroup>
   <ItemGroup>
@@ -94,6 +100,10 @@
       <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Include="Views\NotificationsShell.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="Views\NotificationsView.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
