+++ src/Smallify.Module.Notifications/ViewModels/INotificationsViewModel.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.ObjectModel;
+
+namespace Smallify.Module.Notifications.ViewModels
+{
+	public interface INotificationsViewModel
+	{
+		ObservableCollection<string> Notifications { get; set; }
+	}
+}
