+++ src/Smallify.Module.Notifications/ViewModels/NotificationsShellViewModel.cs
@@ -1,8 +1,5 @@
 ﻿using Prism.Commands;
-using Prism.Events;
 using Prism.Mvvm;
-using Smallify.Module.Core.Events.Notifications;
-using System.Collections.ObjectModel;
 using System.Windows;
 using System.Windows.Input;
 
@@ -10,22 +7,15 @@ namespace Smallify.Module.Notifications.ViewModels
 {
 	public class NotificationsShellViewModel : BindableBase, INotificationsShellViewModel
 	{
-		public NotificationsShellViewModel(IEventAggregator eventAggregator)
+		public NotificationsShellViewModel(INotificationsViewModel notificationsViewModel)
 		{
-			Notifications = new ObservableCollection<string>();
-
 			ExitCommand = new DelegateCommand<Window>((window) => window.Close());
 
-			eventAggregator.GetEvent<NewNotificationEvent>().Subscribe(NewNotificationReceived);
+			NotificationsViewModel = notificationsViewModel;
 		}
 
 		public ICommand ExitCommand { get; private set; }
 
-		public ObservableCollection<string> Notifications { get; set; }
-
-		private void NewNotificationReceived(string notification)
-		{
-			Notifications.Add(notification);
-		}
+		public INotificationsViewModel NotificationsViewModel { get; private set; }
 	}
 }
