+++ src/Smallify.Module.Notifications/Views/NotificationsView.xaml
@@ -1,0 +1,22 @@
+﻿<UserControl x:Class="Smallify.Module.Notifications.Views.NotificationsView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:local="clr-namespace:Smallify.Module.Notifications.Views"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:samples="clr-namespace:Smallify.Module.Notifications.Samples"
+             d:DataContext="{d:DesignInstance Type=samples:NotificationsViewModelData,
+                                              IsDesignTimeCreatable=True}"
+             d:DesignHeight="450"
+             d:DesignWidth="800"
+             mc:Ignorable="d">
+    <Border>
+        <ItemsControl ItemsSource="{Binding Path=Notifications}">
+            <ItemsControl.ItemTemplate>
+                <DataTemplate>
+                    <TextBlock Text="{Binding Path=.}" />
+                </DataTemplate>
+            </ItemsControl.ItemTemplate>
+        </ItemsControl>
+    </Border>
+</UserControl>
