+++ src/Smallify.Module.Notifications/Views/NotificationsView.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Smallify.Module.Notifications.Views
+{
+	/// <summary>
+	/// Interaction logic for NotificationsView.xaml
+	/// </summary>
+	public partial class NotificationsView : UserControl
+	{
+		public NotificationsView()
+		{
+			InitializeComponent();
+		}
+	}
+}
